+++ tests/unit/test_tasks/test_psfgen_resegment/00-01-000_psfgen-build.tcl
@@ -1,5 +1,5 @@
 # pestifer.scripters: 00-01-000_psfgen-build.tcl
-####################### Created Wed Aug 19 09:53:20 2026 #######################
+######################## pestifer 3.18.0  seed 27021972 ########################
 package require PestiferCRot
 namespace import PestiferCRot::*
 package require psfgen